server/engine: Wait for exclusive access to transaction before commit
If anything in our query engine defers execution, there will be
references to the transaction when we reach
`commit_transaction_static()`.  Therefore, change the function to *wait*
for exclusive access to the transaction before attemting to commit.

Note: we now have a potential deadlock if endpoint response body stream
uses transactions because we wait *before* we even give control over to
Hyper that reads from the stream.

Fixes #851