server/deno: Ensure correct transaction on commit
There is no absolute guarantee that the request context we clear is the
one that contains the transaction of our commit because current context
is only restored on `RequestFuture::poll()` but not across all possible
suspension points.

However, we already know in `run_js()` what the correct transaction for
this endpoint invocation is because we start the transaction. Therefore,
stop usign the transaction from `clear_current_context` in favor of the
local variable `transaction` in `run_js()`.