Store an Option is CURRENT_CONTEXT
This asserts that we never try to use a context that is not set. There
are now only two functions that manipulate CURRENT_CONTEXT:
with_current_context and with_new_context. Only with_new_context
mutates it and it always restores the old value.

This should hopefully make sure that we always get the correct
context.