api/chisel.ts: Rename ChiselEntity to Entity
The `Chisel` prefix is redudant because application code has to
explicitly import the class to use it. Furthermore, in the case of a
naming clash (application code also has "Entity" class), applications
can use the "as" keyword to rename the imported class.

So let's rename `ChiselEntity` to `Entity`, but leave a backward
compatible alias class, which is marked as deprecated. Please note that
type aliases *do not* work, so `ChiselEntity` extends `Entity`.