api/chisel.ts: Rename ChiselCursor to Enumerable
The "cursor" name refers to a pointer-like thing, but in our case, it
actually represents a deferred execution of a query. Let's rename the
class to `Enumerable` (terminology borrowed from LINQ), which is more
accurate (although more generic as well) description of what the class
represents. While at it, rename `Entity.cursor()` to
`Entity.enumerate()`.

Leave `ChiselCursor` and `Entity.cursor()` as backward compatible name
too, but mark it as deprecated.