Implement close for QueryStreamResource
I don't seem to be able to read the javascript spec, and can't find
any less formal document saying so, but testing with

    function* foo () {
        try {
            for (let i = 0; i < 3; i += 1) {
                yield i;
            }
        } finally {
            console.log("finally");
        }
    }
    function bar() {
        for (let i of foo()) {
            console.log(i);
            return
        }
    }
    bar();

shows that finally is used. The probably we have is really in the
resource table, but not as obvious as I thought. One thing that is
missing is an implementation of close. This adds a bare minimum that
drop the stream. It should also cancel futures, but this is a start.