remove counter test
I was made aware today that this test exists, but I am confident it
doesn't really work: it works if all requests land in the same thread.
That is true for tests, but won't be true for real production endpoints.

I personally don't think we should support that. It's too much work for
too little gain, and we can trivially support a use case like that with
a model definition.

Happy to discuss in the PR if there is disagreement.