Keep a single module loader for all webworkers
Doing this mostly because it simplifies compiling endpoints from
webworkers. With a single module loader compile_endpoint can insert
the data and tell the webworker to load it. With one module loader per
worker, compile_endpoint would have to pass the data to the endpoint.

We also only load modules when endpoints are created, so this should
not be a serious contention. In fact, it probably makes sense to have
a single module loader for the entire process.