Convert some rust code into TS
In order to handle request in web workers, we have to pass information
down to javascript in a format that can be sent to workers. For
example, we can path a string, but not a function.

This PR converts some of our implementation from rust to TS so that it
can be moved into a web worker in a followup PR.