Add new secrets decryption protocol
Secrets are now encrypted using AES-GCM and we use RSA-OEAP to encrypt
the random key and nonce used on the AES. We do this because AES can
encrypt secrets of arbitrary size efficiently. We keep RSA so we don't
need coordination between the parties that encrypt and decrypt. To
encrypt secrets, all you need is a public key.

Algorithms were picked based on the availability of secure
implementations both in Rust and in the JS ecossystem.

We keep backwards compatibility by trying to parse the data in the new
format (AesPaylaod), and falling back to the old protocol in case this
does not work.