improve crud generator
The first big problem is the JSDoc: `below` doesn't mean anything,
because users using IDEs are not looking at the file, just the snippet.

We need to document it better even, especially because this, and not the
top-level crud, is the main entry point users are looking for.

However, the fact that we need to pass the pathname is annoying,
error-prone, user-unfriendly, hard to document, and quite frankly,
borderline unacceptable since it always has to include (and in the
simple crud version fully match) the name of the file in a product
that aims a being user-friendly.

We add a deno function that exports the current path to the Typescript
layer, and then use that to compose the template. The basic crud()
function now needs nothing, and the version with more setup has to pass
just the suffix.

(cherry picked from commit c65a01a48ef48bd3757e4c82466917ab8fb38680)