Don't try to update submodule from build.rs
Cargo has no notion of short circuiting a build. If a build.rs is run,
all its dependencies are built. Given that, it is critical that a
build.rs prints cargo:rerun-if-changed=PATH, but we don't have a file
that we can print in here. We would like for this to run when
submodules change, but that info is in the guts of a git pack file.

Users that want it easy should just set submodule.recurse in the git
config.