Introspection support with OpenAPI
This adds introspection support to `chiseld` by exporting a
`__chiselstrike/introspect` endpoint that returns OpenAPI 2.0 compatible
description of the endpoints:

```console
% curl localhost:8080/__chiselstrike/introspect
{
  "swagger": "2.0",
  "info": {
    "title": "ChiselStrike API",
    "version": "0.0.0"
  },
  "paths": {
    "/__chiselstrike/auth/callback": {},
    "/__chiselstrike/auth/user/": {},
    "/__chiselstrike/introspect": {},
    "/dev/hello": {}
  },
  "definitions": {}
}%
```

Note #1: The OpenAPI definition doesn't have HTTP methods because we
don't have any mechanism to know what HTTP methods are supported. This
causes Postman, for example, not not show any paths. We need to address
this as a follow-up work at some point.

Note #2: I picked OpenAPI 2.0 because there are existing Rust libraries for
serializing that format. I assume it's supported by most tools, but I
verified that it's compatible with Postman.