Add a worker skeleton
This includes the logic for sending only one request at a time, but
not much more.

Given the restriction that one cannot call from JS to Rust and back to
JS, I think I will now have to do a dance at worker creation time to
setup a channel between the two rust threads. Without that I would
have to convert way too much code to JS, since anything in rust asking
questions about types or secrets would have to be in JS.

This skeleton will be used for all messages between the main thread
and the workers, including the above dance.