Move work to Web Workers
This regresses our ability to handle streaming, but fixes the
concurrency problems.

For now it includes only one worker, so requests are serialized.

I think this can be simplified to avoid so much chit-chat between the
worker and the main thread. It should also not be too hard to bring
back streaming or add more workers.

Given how pressing the concurrency problems are, I think this is worth
it.