prevent spam of messages when reading secrets
A rustls issue is causing our logs to be spammed. There is nothing we
can do because the print comes from rustls itself. There is an issue
for that with rustls
(https://githubhot.com/repo/ctz/rustls/issues/1012), and other reporters
confirmed that pinning rustls to 0.20.2 fixes this (our previous version
was 0.20.4, and 0.20.3 is also affected).

Fixes #1064