Add ChiselRequest
This is a wrapper around Request, that allow us to add fields and
methods that may be useful for ChiselStrike users. Those methods
can certainly be obtained in other ways, likethe currently logged in
user, but having them in the header leads to a lot less typing.

Two are of particular interest: the pathParameters, since the user can
always get it from the URL, but then the logic to remove the endpoint
path and the version are pushed to the user. And as mentioned, the
currently logged in user.

Fixes #998