cli: Perform apply only on changed files
Track which files actually got changed when performing apply. This is an
optimization, but also needed for correctness for the next patch. That
patch makes "chisel dev" apply when file is renamed, but that generates
two filesystem events: for the old file and the new file, but we only
want to perform apply for the latter.