chiselstrike-api: Drop yarn.lock file
There's two build tools in the Node ecosystem: npm and yarn. The former
generates a "package-lock.json" file to specify exact dependencies
whereas the latter generates a "yarn.lock" file. There's absolutely no
reason to have *both* of the lock files in the source tree. As we are
using "npm" in our scripts, let's just consolidate to using the
"package-lock.json" file.