Move rollback control to javascript
In the included testcase we now rollback when we didn't before. I am
unable to detect this in a test with the current asserts we have, but
I can confirm it by just logging when rollbacks and commits happen.

This still don't always rollback, but I think it is a step in the
right direction.