Enhance OpenAPI with version and application name
On our platform, recently the Profusion guys resorted to creating a
special endpoint with a string that they manually changed every time
they changed the code, to make sure that the deployed version was
using the right code.

Needless to say, that's ridiculous and we should have a better way.
This patch grabs information from git - if available - and package.json
- if available - to construct meaningful strings that can be queried
  from the introspect endpoint.