sanitize api_version
We will use the api version as part of the URL path, so we should
disable things that will have potential for mayhem.

In particular, because early on we disallowed dots and used dots to
compose the type names, api versions with dots effectively corrupt our
database: we hit an assert (thankfully, could be worse) on the read
side.