Use NextAuthUser instead of OAuthUser
Instead of using the ChiselStrikeToken header, require the client to
provide a NextAuthUser ID in the ChiselUID header.  The NextAuthUser
row will have been stored separately by the NextAuth DB adapter.

Delete the mechanism of storing OAuthUser rows and issuing tokens.
Also delete the endpoint previously used to obtain the username from
the ChiselStrikeToken value; now the frontend gets the current user's
info from the NextAuth session.

This is the first step of a big change; other steps are coming in
subsequent commits.  In particular:

- OAuthUser isn't deleted yet, just unused

- the new way of finding the currently logged-in user hasn't been
  plumbed through to TypeScript yet; tests relying on the old way have
  been temporarily deleted

- the endpoint handling OAuth callback isn't deleted yet

Signed-off-by: Dejan Mircevski <dejan@chiselstrike.com>