cli: Fix disappearing endpoints with 'chisel dev'
One more fix in the never-ending saga of 'chisel dev' breakage...

The story starts with 'chisel dev' not triggering on file rename
(discussed in #906). To fix that, there was an intermediate commit
67659f06 ("cli: Perform apply only on changed files") that attempted
to address the issue of some editors (hello vim!) generating temporary
files that get renamed on save, which resulted in models being imported
multiple times (for the old and new name with renames).

However, commit 67659f06 pushed the logic of ignoring temporary files at
the wrong layer, resulting in models being first dropped, and then
re-added causing errors (discussed in #1138). That bug got fixed in
commit b3f09430 ("cli: Don't pass temporary files to apply()"), but that
revealed that the optimization was incorrect.

The reason we cannot filter paths in apply() is because none of the
`EndPointCreationRequest` stuff is actually used (it's dead code).
Instead, we always pass a full set of model and endpoint definitions to
the server. So if we filter a path, then the server thinks that path no
longer exists, and removes it.

Fix disappearing endpoints with 'chisel dev' by undoing the broken
optimization in commit 67659f06.

Fixes #1156

(cherry picked from commit 304479e0992f9b4238b2b6c26282c25fa918fb68)