Handle closed queries
Deno's future cancellation works by calling the wakers, so there is no
guarantee as to when the futures are dropped. We need something more
immediate, so use a weak pointer.

The net result is that now we get a log in chiseld similar to:

  Unhandled exception: Error: Promise rejected without a handler: Error: Closed resource
      at file:///worker.js:5:11
      at Function.findAll (file:///chisel.ts:342:35)

(cherry picked from commit 1033e76772fe0c83055ea3f69e802d44bf2e87d7)