Handle exceptions when resolving promises
Deno doesn't map exceptions to Result. Since it doesn't, exceptions
getting to the rust interface are unhandled, are added to
pending_promise_exceptions and cause an unrelated future to fail.

Solve the problem by always handling promises and wrapping produced
value or error and then mapping that to a Result.