Add introspection routes to the root of the api version
As Michael correctly pointed out, the __chiselstrike/introspect endpoint
is not really discoverable, as you would essentially had to guess it.

It is best to add introspection to the root of the api_version. It also
makes it more natural to have an endpoint that shows all routes, from
all versions.

Because the test was not merged before, we merge it now.