Move docs to their own repository
We currently have to update two repositories to publish documentation
changes. To make our lives easier, let's move documentation to its own
repository, which we can make a Vercel project:

https://github.com/chiselstrike/docs.chiselstrike.com

(Full git history is preserved.)

After this change, we no longer have to manually copy any changes over.
Instead, the workflow will be as follows:

  1. Merge a documentation change to `main`
  2. Cherry-pick or merge a change to `deploy` and push

That's it!

Note! In the future, when we need versioning, it's anyway going to be
file-based so there's no benefit in keeping source code in the source
tree:

https://docusaurus.io/docs/versioning