merge meta and query dbs
Users that were using both are automatically migrated. And in the
process, we move the new db to a dot-file, to keep users' workspaces
clean.

We migrate the files from the locations the users specified. So users
who specified a non-default location will get them migrated - but new
files will not be created.

We'll follow a deprecation schedule and eventually remove these options.