Keep a map of bodyParts
There are two missing features with the web worker based
implementation that we want to add:

* Streaming of results
* Multiple workers

The first one means that we can't wait to receive all the parts before
returning. The second one means that endpoint.ts has to track
concurrent partial bodies.

This patch prepares for both by changing bodyParts from an array to a
map of arrays. With that we don't have to copy bodyParts since a new
request doesn't interfere with the old one and we already have a map
in place for when we add multiple workers.