Don't buffer output in curl
Most tests has small output where this doesn't make a difference, but
for testing streaming I noticed that curl was downloading a lot of
data, even when the output was piped to head.

Without buffering curl feeds output directly to the output and the
test runs much faster.

We could add -N only to tests that need it, but it seems better to
have it here since I don't expect any test that will benefit from
buffering in curl.