Relax handling of rejected promises without handles
Given

  Deno.core.setPromiseRejectCallback((type, promise, reason) => {
      console.log(type, promise, reason);
  });
  async function bar() {
      throw "abc";
  }
  async function zed() {
      try {
          await bar();
      } catch {
      }
  }
  await zed();

V8 will first call the callback with PromiseRejectWithNoHandler (0) to
let us know that a promise was rejected and then call it again with
PromiseHandlerAddedAfterReject (1) to let us know that that promise
now has a handle and we don't need to worry about it.

This means that we cannot just abort when the callback is first
called, we have to wait to see if it is called again with
PromiseHandlerAddedAfterReject. Unfortunately we don't know how long
to wait, so this patch adds a check during shutdown only.