Add the endpoint_compile crate
This small crate refactors the special logic for compiling endpoints,
as opposed to plain ts files, to a new crate. This cannot be part of
tsc_compile sice tsc_compile is used to compile api which is used by
the endpoints.

With this my_tsc complies endpoints the same way that "chisel apply"
does.