Move build code to a new package
As far as I can tell, the only way in rust to share code between
build.rs and src/lib.rs is to move it to another package and it both
as a plain dependency and as a build dependency.

Since we want to be able to use tsc_compile without a snapshot, move
code that will be needed in lib.rs from build.rs to another package.