Output compiled import files
With this we output the javascript corresponding to imported
typescript files.

The fix is done by changing write_impl to record more files and having
a post processing step that maps JS files to the TS files they came
from.

It is possible to do this in write_impl itself, but the post
processing version is simpler.

A followup PR will make write_impl trivial by doing all the work on
the post processing step.