cli/apply: Fix chiselc in Node mode
We currently transform the "bundler entry file" with `chiselc`, but it's
not the right file to transform. Instead, we need to transform the
actual endpoint file.

Therefore, explicitly transform endpoints before generating the bundler
entry files, and then drop `chiselc` from the bundler step altogether.

Please note that we first transform all endpoints concurrently before
even starting to run the bundler. This ensures that we always have the
file transformed before bundling. However, we can make this faster by
starting the bundler immediately when `chiselc` completes. (I did try
this approach, but it turned to be an adventure of its own, so I am
deferring it for later.)

Fixes #1317