Return a JsonObject from getSecret
Also use a simple comparison to check if we have a new one. We already
have a copy of the secrets in each worker, so using a hash was not
saving much memory. getSecret was also already decripting the secrets,
so avoiding a json parse was also not saving that much cpu.

With this all errors getting secrets are handled uniformly.