crud.rs: (almost) Always provide next/prev page.
This commit changes the behavior of paging so that we now provide
both prev/next page for all non-empty pages. The old code tried
to detect if we are on first/last page and not provide link to a page
that is potentially empty. This is problematic as with dynamic additions,
we can never perfectly estimate whether the next/previous page will be
empty so we as well might not even try and thus simplify the code.