Use upstream deno URL
We don't need any patches for deno at the moment. We are getting
security warnings from github because the main branch in our fork of
deno is old (which is OK, since we use tags, not the main branch).

With this we should be able to delete the fork. It is easy enough to
create it again if we ever need to.