execute inserts instead of fetch_one
Although there is a returning statement in the insert statement, we
are currently not assigning the end result to anything. That indicates
this is an appendage that is no longer needed.

Also, for some unexplained reason that can easily be an sqlx issue,
fetch_one fails in some queries, so the move to execute has an extra
advantage.

(Accidentaly) Fixes #1373