Fix clearing secrets
If reading secrets during startup, the last_secrets variable had the
wrong value since it was not aware of the initial secrets. This means
we would never clear the secrets, since last_secrets already thought
they were clear.

This simply removes the attempt at optimizing the secret reading
logic. We can add it back, but in read_secrets, where we can track all
reads, startup and periodic. Before adding it back I would like to see
if it this is actually expensive.