Fix some type assertions
This is mostly Jan Plhak's work, I just simplified it a bit.

This changes the meaning of Operator<T> from something that produces
Ts to something that reads Ts. Neither is correct, but this at least
removes the most broken type assertion.

To make this sound we have to somehow map

class Foo extends Chisel.ChiselEntity {
  bar: string = "";
  zed: number = 0;
}

To

type Foo2 = {bar: string, zed: number}

Which is what is actually read from the DB. I don't see how to do that
with pure typescript.

To avoid all type assertions but the one when we actually read from
the DB, we need to somehow encode the op chain at the type system, so
that each recursive operation knows what type it consumes and
produces.