Change the code_map urls we use
Currently an endpoint /api_ver/foo gets its source stored in a
hashtable keyed with the url
file:///api_ver/foo.js?ver=<file_version>.

Since that same hash table will have to store modules that are not
endpoints, this PR changes the url to
file:///api_ver/endpoints/foo.js?ver=<file_version>.

Which can also be used to store models or other files imported by the
endpoints.

There is a small hack in that we still store in the database using the
old format. This will be changed in a followup PR since that part
requires a migration.