watch the base directory only when in dev mode
We try to watch the policies, models and endpoints directories for
change, but that's not ideal.

First, from the crate's documentation:
```
On some platforms, if the path is renamed or removed while being
watched, behaviour may be unexpected. See discussions in 165 and 166.
If less surprising behaviour is wanted one may non-recursively watch the
parent directory as well and manage related events.
```

On top of that, if some of those directories don't exist when we
first invoke dev, watch will error out and `chisel dev` will fail (#1420)

Since those directories can be created later, Let's follow the crate's
suggestion and watch the entire directory, filtering out stuff we don't
want.

Fixes #1420