Fix FIXME and cast
The FIXME was wrong. When we have a ColumnsSelect, T involves a
Pick<AnEntity, ...>. This means that it is a simple record and rust
has set all the properties.

The cast was wrong when there was no ColumnsSelect. In that case, T is
a base of ChiselEntity, which rust did not construct. We will
construct it using ctor.