Split evalOpsRecursive among the Operators
Thanks a lot to Jan Plhak for the suggestion. TypeScript has to
associated types, so as far as I can tell this is the only way to make
eval more type safe.

The type assertion now is moved to the one place it is needed and it
is right after the call to rust, which is where it is reasonable to
need it.