Make type more flexible
The BodyInit in lib.dom.d.ts accepts any ReadableStream. The one in
lib.deno_fetch.d.ts only ReadableStream<Uint8Array>.

This changes the code to use ReadableStream<Uint8Array> so that it
works with both. On the rust side, op_chisel_read_body returns
ZeroCopyBuf, so this is reasonable.