Use a super long timeout
We have been hitting timeouts in the bots and I have seen them locally
too.

Some operations, like collecting backtraces in debug builds, are
super slow, so it is not clear if we have a bug causes the server to
be stuck or we just need more time when the system is under load.

An indication that we might just need more time is that I never get a
timeout when running a test in isolation.