migrate lazy_static to one_cell
The rationale behind that is that once_cell is more idiomatic and easier to use than lazy_static, and doesn't require the use of macros. On top of that, there is an effort to introduce it into std: https://doc.rust-lang.org/std/lazy/struct.Lazy.html.

We already transitively depended on this crate, so this does not impact compile time.

lazy_static was also hiding unused variables.