Deepkit demo
This is a hack, but should let others play with deepkit's type
reflection. This still uses a patched deepkit, but at least the logic
for bundling it is in a build.rs.

See cli/tests/lit/reflection.deno for an example of how to use it. If
you are curious about the magic, checkout the output produced by
my_tsc on tsc_compile/tests/goldilocks.ts.