run all endpoints through the same esbuild invocation
we are invoking one esbuild command per endpoint. However esbuild
would prefer to be called once with all files, being passed a directory
instead.

This could be seen as just a minor performance optimization, but as it
turns out, if we're really unlucky, it turns out that esbuild *really*
doesn't like to be called in parallel

Fixes #1526