api improvements: allow users to return arbitrary types
We mandate a Response to be returned, meaning that users have to do
returnFromJson, or in other ways construct responses.

However in many cases, specialized endpoints want to return well
defined things like an array of models, a model, or even strings.

Allow that to happen, by converting to a Response internally.
This is a major UX boost that allow us to craft much shorter endpoints