Improve error messages
I encountered these messages when I was trying to import 'memfs' and
'isomorphic-git' and applying some hacks.

Do not use error messages in gerund: when we do .context("doing X") on
an existing error, the resulting error says "doing X" and the original
error is stored as a _cause_. The resulting error message becomes very
confusing: you either see something like "parsing endpoints; Caused by
Y" or just "parsing endpoints".

The messages in .context() should instead form a proper error
message that is understandable on its own, so it should say "could
not do X" or "failed to do X".