api: adds the map operator
map is a classic operation that is provided by Array and frankly many
other primitives.

This simple version executes the code in typescript, but already has
the added advantage of not forcing the full materialization of the array
prior to map, as would be the case with toArray.map().

In the near future we can pass many of these functions to the database.