make sure paths are respected within .gen
There's currently a pretty serious problem with chiselc and endpoints
within subdirectories. The algorithm to figure out the name of the
generated file is to just take the filename() component of the endpoint
path, which has two big problems:

1. endpoints/foo/bar.ts will override endpoints/bar.ts in .gen.
2. import paths containing "../../" will not work.

The best way forward, especially because of relative paths for imports,
is to keep path structures the same, just replacing endpoints with .gen.

We'll have issues if the user uses either absolute paths or ./ as paths
in the manifest, but those are likely bad ideas (with the first
potentially being a security issue in the managed deployments) anyway.
So this patch also prevents them by checking early and making sure those
things are not present.

Fixes #1511