allow arbitrary defaults
This PR allow us to write things like:

  a: number = Date.now()

in defaults.

We need surprisingly little to make this work, which makes me very
happy. Essentially, Entity.build() will already execute those functions,
so it works just fine.

We just need to make sure that the schema evolution works, which also
happen to work just fine: for complex expressions, we just treat them
as "no defaults". That makes sense, because if you have something like
"Date.now()", the behavior the user expects for things that didn't
exist before is not to retrofill with the current date...

So we allow adding mandatory types with literals and change the error
messages accordingly. Complex expressions behave like no-default.

We do miss the opportunity to allow things like "string".length and "1 +
2", which are not literals, but are pure and could be executed compile
time. However this is good for now, and we could in the future expand
chiselc to allow for those expressions to be parsed as literal defaults.