allow deletion of empty models
Currently we disallow models to be removed from files at all, because
that can destroy data in the database. Deletion has to happen with
an apply flag.

This is very inconvenient, though, for both local development, where a
user may be just experimenting with types, or hosted chiseltrike, where
apply is automatically done through git, so there is no way to pass
flags.

We should be more tolerant, and allow models to be deleted if they are
empty.