server: Make '--inspect-brk' pause once at startup
Currently, the '--inspect-brk' flag makes ChiselStrike runtime pause and
wait for a debugger for every endpoint invocation, which makes the
setting unusable for things like heap profiling.

So let's change the semantics of '--inspect-brk' match what Deno does
(the CLI has the same flag): pause once at startup to wait for a
debugger to attach, but then keep on running.

Note that users can still get the per endpoint behavior if they set a
breakpoint when a debugger starts up.