granular readiness
We always return 200 on the readiness probe, but there are some race
conditions here, due to which it is still possible for the readiness
internal health endpoint to say we are ready to take requests when
in reality we are not.

This increases the granularity and gives us more control over how
we declare whether or not we're ready.