CRUD: add a PATCH verb
This is crucial for objects with non-literal defaults. Consider for
example, the following object:

```
export class Joke extends ChiselEntity {
        createdAt : number = Date.now()
        name : string
        content : string
}
```

because "createdAt" has a default, it is convenient not to specify it
when creating the object. However, if you call a CRUD PUT on an existing
object, the whole object gets recreated and have the default it had
before (in this case, a creation time).

With a PATCH, we can pass just name and content, and createdAt will be
left alone.

Fixes #1213