policies in typescript
This PR implements the bare minimum of complex policies in typescript.
In particular:
* policies are not persisted
* only access control works
* only create and read are checked
* no distinction between create and update yet
* doesn't work with crud() reads
* doesn't work with delete
* the audit log is just a console print, this should go into a
  __chiselstrike endpoint and/or kafka
* I don't lint anything to prevent loops or other things we want
  to disallow

I tested this with the following file:

```
import { Access, ActionResult } from "@chiselstrike/api"
import { User } from "../models/User"

Access(User, {
    read: (x, req) => { return ActionResult.Allow },
    POST: (x, req) => {
        if (x.name == "Glauber") {
		return ActionResult.Log
	} else {
		return ActionResult.Deny
        }
    }
});
```

This is quite simplistic, and have the Model as the entrypoint. So
it is hard to write policies that take many models into account.
But we have to start somewhere. We may want to make the functions
asynchronous, which would allow us to issue findOnes() to the other
models.

The obvious next step is to pass information about the request,
and then we can check for token in headers, logged in user, etc.

I have the following functions planned for the near future:

```
setLocality(User => {
    if x.country == "eu" {
        return Jurisdiction.EU
    } else {
	return Jurisdiction.Any
    }
})

setEncryption(User, Encryption.Rest)
setEncryption(BlogPost, Encryption.Tls)
```