Remove the word "endpoint" from public interface
PR #1618 replaces the concept of "endpoint" with the concept of a
"route", so let's change this now, before the main PR lands.

Notably, this changes the default directory "endpoints/" to "routes/".
Unfortunately, even though the directory appears to be configurable in
"Chisel.toml", it is in fact hard-coded in several places, so the change
in this commit is *not* backward compatible. Next commit will try to
hack around this, so that we retain backward compatibility.

This change also affects some error messages, which are brought in
alignment with #1618 by this commit.