Keep backward compatiblity with "endpoints/" directory
The directory with routes/endpoints is in fact hard-coded multiple
times, even though "Chisel.toml" pretends that it is configurable. This
commit adds hard-coded support for the old "endpoints" directory along
with the new "routes" directory, to ensure that we remain backward
compatible even if we switch the defalt to "routes".