Remove the `Chisel` global variable
In modern JavaScript, it is usually better to use imports rather than a
global variable. For example, to type-check code that uses the global
variable, we must feed `tsc` with an extra file that specifies the type
of the variable. Getting rid of the `Chisel` variable will simplify our
code in the fugure.