Introduce --debug flag to return internal server errors
On production, we should not return any details about internal server
errors, because they might be useful to potential attackers. But it is
very convenient to return the errors while developing or in tests, so
introduce a new option `--debug`, which is enabled in tests and in
`chisel dev` and which keeps the original behavior of returning the
error.