api: Switch ChiselEvent to use Blob type
We need to persist key and value to a database to make event publishing
transactional. To do that, let's switch the `ChiselEvent` to use the
`Blob` type, which is a nicer type for an ORM.

As a bonus, it becomes easier to consume events as JSON as illustrated
in the docs.