create-chiselstrike-app: Generate Dockerfile
This adds Dockerfile generation to `create-chiselstrike-app`, which
allows developers to build an image of their application and its
dependencies, including ChiselStrike, to easily deploy to applications
to Kubernetes, Fly.io, and other services.

Please note that the Dockerfile only exports port 8080, which means that
there's no access to the ChiselStrike RPC from outside. We can re-visit
the decision once we have TLS support for gRPC:

https://github.com/chiselstrike/chiselstrike/issues/1782

Also note that the generated image is pretty big because we need to run
`npm install` to fetch dependencies, which requires Node to be present.
The `chisel apply` command also has a hard-coded dependency to `npx`.

Also, we're currently unable to use Alpine Linux, for example, as a
super-slim base image because of:

https://github.com/chiselstrike/chiselstrike/issues/1781