Use serde_yaml to parse policy files
Simplifies our code while also rejecting misspelled YAML elements that
we were silently ignoring before.  This avoids running broken policies
without warning.

One wrinkle: when serde_yaml sees a struct member of type String, it
parses the corresponding YAML value as a string even if that value is a
number or a Boolean.  So `name: 1234` will be accepted as if it said
`name: "1234"`.  This seems acceptable, especially since IDK how to
change it.  But I had to disable some testcases because of it.

Signed-off-by: Dejan Mircevski <dejan@chiselstrike.com>