ChiselStrike adapter for NextAuth.js
This adds a ChiselStrike adapter for NextAuth.js. The adapter is based
on the work Dejan did earlier:

https://github.com/nextauthjs/next-auth/pull/4573

The package is standalone and, therefore, does not require NextAuth.js
to integrate our work into their tree for our users to use this.

Please note that although you can't run the tests in the tree, they do
pass if you integrate the package with the NextAuth.js repository and
enable the "workspace" tagged dependencies (that only pnpm understands).