server/kafka: Don't kill server on receive failure
It's easy to trigger the receive failure path. For example, add a topic
handler to the `events` directory for a non-existing topic. So let's not
kill the server when that happens, but instead, warn about it, which is
better behavior for transient falures as well.

Spotted by the `test_kafka_apply` test case that -- for some reason --
started failing on my machine.