improve Dockerfile for project template
There are many platforms around where you may want to deploy
ChiselStrike where there is no local volume (one real example:
digital ocean). For those, we have to pass an environment variable
with a connection string for an external database.

The ARG statement for Dockerfile allows us to have a default, so we
can default to a local image, meaning it will still work when no
arguments are passed. But this opens the door for users to use
external databases more easily.